Backend/src/requirements.txt
mysql-connector-python
python-dotenv
PyJWT
bcrypt